+++ src/main/java/core/basesyntax/strategy/impl/DefaultDiscountService.java
@@ -3,7 +3,7 @@ package core.basesyntax.strategy.impl;
 import core.basesyntax.strategy.DiscountService;
 
 public class DefaultDiscountService implements DiscountService {
-    private static final int DISCOUNT = 0;
+    private static final double DISCOUNT = 0;
 
     @Override
     public double getDiscount() {
