+++ src/main/java/core/basesyntax/strategy/impl/NewYearDiscountService.java
@@ -3,7 +3,7 @@ package core.basesyntax.strategy.impl;
 import core.basesyntax.strategy.DiscountService;
 
 public class NewYearDiscountService implements DiscountService {
-    private static final int DISCOUNT = 20;
+    private static final double DISCOUNT = 20;
 
     @Override
     public double getDiscount() {
